Flatten valueMap
