+++ src/util/path.js
@@ -1,0 +1,7 @@
+export const split = path => /(\[\d+\])|\./[Symbol.split](path).filter(Boolean)
+
+export const join = (...parts) =>
+  parts.reduce((path, part) => {
+    const prefix = path && part ? "." : ""
+    return `${path}${prefix}${part}`
+  }, "")
